test: add ts-jest (closes #3171)
