+++ jest.config.js
@@ -1,5 +1,8 @@
 /* eslint-env node */
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
 module.exports = {
+  preset: 'ts-jest',
+
   testEnvironment: 'jsdom',
   roots: ['<rootDir>/test-jest'],
   maxWorkers: '50%',
