+++ package.json
@@ -86,7 +86,7 @@
     "@types/d3-interpolate": "^3.0.1",
     "@types/d3-scale": "^4.0.2",
     "@types/d3-shape": "^3.1.0",
-    "@types/jest": "^26.0.24",
+    "@types/jest": "^27.5.1",
     "@types/lodash": "^4.14.144",
     "@types/node": "^14.18.34",
     "@types/react": "^16.0.0",
@@ -134,6 +134,7 @@
     "rimraf": "^3.0.2",
     "sinon": "^9.2.2",
     "terser-webpack-plugin": "^5.1.1",
+    "ts-jest": "^27.1.5",
     "ts-loader": "^8.0.12",
     "typescript": "^4.1.3",
     "update-browserslist-db": "^1.0.10",
